added live loop
