+++ src/Main.java
@@ -15,7 +15,6 @@ public class Main {
     public static void main(String[] args) {
         SessionFactoryMaker.getFactory();
         System.out.println("DB connection successful!");
-        fillDb();
         runProgram();
     }
 
@@ -28,6 +27,8 @@ public class Main {
                 String input = sc.nextLine();
                 switch (input) {
                     case "1" -> Question.add(sc);
+                    case "2" -> CommandHandler.takeExam(sc);
+                    case "3" -> CommandHandler.fillDb();
                     case "0" -> runProgram = false;
                     default -> System.out.println("Incorrect input! Try again.");
                 }
@@ -37,92 +38,4 @@ public class Main {
         }
     }
 
-    public static void fillDb() {
-        Random r = new Random();
-
-        SessionFactory sessionFactory = SessionFactoryMaker.getFactory();
-        Session session = sessionFactory.openSession();
-        session.getTransaction().begin();
-
-        // Crete students
-        Student student1 = new Student();
-        student1.setName("Laurynas");
-        Student student2 = new Student();
-        student2.setName("Jonas");
-        Student student3 = new Student();
-        student3.setName("Mantas");
-        session.persist(student1);
-        session.persist(student2);
-        session.persist(student3);
-
-        // Create exams
-        Exam exam1 = new Exam();
-        exam1.setName("CS");
-
-        Exam exam2 = new Exam();
-        exam2.setName("Math");
-
-        Exam exam3 = new Exam();
-        exam3.setName("English");
-
-
-        // Questions for exams
-        for (int i = 0; i < 3; i++) {
-            Question question  = new Question();
-            question.setQuestion("Question " + i);
-            question.setExam(exam1);
-            question.setAnswerA("answerA");
-            question.setAnswerB("answerB");
-            question.setAnswerC("answerC");
-            question.setAnswerId(r.nextInt(3));
-            session.persist(question);
-        }
-
-        for (int i = 0; i < 3; i++) {
-            Question question  = new Question();
-            question.setQuestion("Question " + i);
-            question.setExam(exam2);
-            question.setAnswerA("answerA");
-            question.setAnswerB("answerB");
-            question.setAnswerC("answerC");
-            question.setAnswerId(r.nextInt(3));
-            session.persist(question);
-        }
-
-        for (int i = 0; i < 3; i++) {
-            Question question  = new Question();
-            question.setQuestion("Question " + i);
-            question.setExam(exam3);
-            question.setAnswerA("answerA");
-            question.setAnswerB("answerB");
-            question.setAnswerC("answerC");
-            question.setAnswerId(r.nextInt(3));
-            session.persist(question);
-        }
-        //Assign exams to students
-        ArrayList<Student> exam1_students = new ArrayList<>();
-        exam1_students.add(student1);
-        exam1_students.add(student2);
-        exam1_students.add(student3);
-
-        ArrayList<Student> exam2_students = new ArrayList<>();
-        exam2_students.add(student1);
-        exam2_students.add(student3);
-
-        ArrayList<Student> exam3_students = new ArrayList<>();
-        exam3_students.add(student2);
-
-
-        exam1.setStudents(exam1_students);
-        exam2.setStudents(exam2_students);
-        exam3.setStudents(exam3_students);
-
-        session.persist(exam1);
-        session.persist(exam2);
-        session.persist(exam3);
-
-        session.getTransaction().commit();
-
-
-    }
 }
